refactor wizard: create ext address
