+++ asistente_ladm_col/gui/wizards/select_features_on_map_wizard.py
@@ -21,7 +21,8 @@
  ***************************************************************************/
  """
 from qgis.PyQt.QtCore import QCoreApplication
-from qgis.PyQt.QtWidgets import QMessageBox
+from qgis.PyQt.QtWidgets import (QMessageBox,
+                                 QPushButton)
 from qgis.core import Qgis
 
 from ...config.general_config import (PLUGIN_NAME,
@@ -41,7 +42,7 @@ class SelectFeaturesOnMapWizard:
         self.canvas.mapToolSet.disconnect(self.map_tool_changed)
         reply = QMessageBox.question(self,
                                      QCoreApplication.translate(self.WIZARD_NAME, "Stop {} creation?").format(self.WIZARD_FEATURE_NAME),
-                                     QCoreApplication.translate(self.WIZARD_NAME,"The map tool is about to change. Do you want to stop creating {}?").format(self.WIZARD_FEATURE_NAME),
+                                     QCoreApplication.translate(self.WIZARD_NAME, "The map tool is about to change. Do you want to stop creating {}?").format(self.WIZARD_FEATURE_NAME),
                                      QMessageBox.Yes, QMessageBox.No)
 
         if reply == QMessageBox.Yes:
@@ -118,3 +119,42 @@ class SelectFeaturesOnMapWizard:
 
     def disconnect_signals_controls_select_features_on_map(self):
         raise NotImplementedError
+
+    def save_created_geometry(self):
+        message = None
+        if self._layers[self.EDITING_LAYER_NAME][LAYER].editBuffer():
+            if len(self._layers[self.EDITING_LAYER_NAME][LAYER].editBuffer().addedFeatures()) == 1:
+                feature = [value for index, value in self._layers[self.EDITING_LAYER_NAME][LAYER].editBuffer().addedFeatures().items()][0]
+                if feature.geometry().isGeosValid():
+                    self.exec_form(self._layers[self.EDITING_LAYER_NAME][LAYER])
+                else:
+                    message = QCoreApplication.translate(self.WIZARD_NAME, "The geometry is invalid. Do you want to return to the edit session?")
+            else:
+                if len(self._layers[self.EDITING_LAYER_NAME][LAYER].editBuffer().addedFeatures()) == 0:
+                    message = QCoreApplication.translate(self.WIZARD_NAME, "No geometry has been created. Do you want to return to the edit session?")
+                else:
+                    message = QCoreApplication.translate(self.WIZARD_NAME, "Several geometries were created but only one was expected. Do you want to return to the edit session?")
+
+        if message:
+            self.show_message_associate_geometry_creation(message)
+
+    def show_message_associate_geometry_creation(self, message):
+        msg = QMessageBox(self)
+        msg.setIcon(QMessageBox.Question)
+        msg.setText(message)
+        msg.setWindowTitle(QCoreApplication.translate(self.WIZARD_NAME, "Continue editing?"))
+        msg.addButton(QPushButton(QCoreApplication.translate(self.WIZARD_NAME, "Yes")), QMessageBox.YesRole)
+        msg.addButton(QPushButton(QCoreApplication.translate(self.WIZARD_NAME, "No, close the wizard")), QMessageBox.NoRole)
+        reply = msg.exec_()
+
+        if reply == 1: # 1 close wizard, 0 yes
+            # stop edition in close_wizard crash qgis
+            if self._layers[self.EDITING_LAYER_NAME][LAYER].isEditable():
+                self._layers[self.EDITING_LAYER_NAME][LAYER].rollBack()
+
+            message = QCoreApplication.translate(self.WIZARD_NAME, "'{}' tool has been closed.").format(
+                self.WIZARD_TOOL_NAME)
+            self.close_wizard(message)
+        else:
+            # Continue creating geometry
+            pass
