+++ asistente_ladm_col/ui/wizards/cadastre/spatial_unit/wiz_create_right_of_way_cadastre.ui
@@ -128,7 +128,7 @@
             </spacer>
            </item>
            <item row="4" column="0" colspan="2">
-            <widget class="QRadioButton" name="rad_digitizing">
+            <widget class="QRadioButton" name="rad_create_manually">
              <property name="text">
               <string>Digitizing Polygon</string>
              </property>
